+++ src/array_subset/iterators/linearised_indices_iterator.rs
@@ -60,6 +60,18 @@ impl LinearisedIndices {
         }
     }
 
+    /// Return the number of indices.
+    #[must_use]
+    pub fn len(&self) -> usize {
+        self.subset.num_elements_usize()
+    }
+
+    /// Returns true if the number of indices is zero.
+    #[must_use]
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
     /// Create a new serial iterator.
     #[must_use]
     pub fn iter(&self) -> LinearisedIndicesIterator<'_> {